safari-534.53-branch/LayoutTests/plugins/window-open.html

<html>
<head>
<script>
    if (window.layoutTestController) {
        layoutTestController.setCanOpenWindows();
        layoutTestController.dumpAsText();
        layoutTestController.waitUntilDone();
     }
     
</script>
<body style="margin:0px;overflow:hidden">
<object name="testPlugin" type="application/x-webkit-test-netscape" testwindowopen="true">
</object>
<div>
    This tests that window.open invoked by a plugin via NPN_Invoke without a javascript context succeeds.
</div>
<body>
</html>
