safari-534.53-branch/LayoutTests/plugins/window-open-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: WINDOW OPEN SUCCESS

This tests that window.open invoked by a plugin via NPN_Invoke without a javascript context succeeds.
